// src/entities/user.entities.d.ts
export declare class User {
    id: number;
    name: string;
    email: string;
    password: string;
    age: number;
    refreshToken?: string;
    resetToken?: string;
    resetTokenExpiry?: Date;
    role: string;
    createdAt: Date;
    updatedAt: Date;
}
//# sourceMappingURL=user.entities.d.ts.map